// apps/help-desk/src/components/generated/models/ticket-model.ts
import type { Company } from './company-model';
import type { Department } from './department-model';

export const TicketPriorityKeyToLabel = {
  'PriorityKey0': 'Low',
  'PriorityKey1': 'Medium',
  'PriorityKey2': 'High',
  'PriorityKey3': 'Critical'
} as const;
export type TicketPriorityKey = keyof typeof TicketPriorityKeyToLabel;

export const TicketStatusKeyToLabel = {
  'StatusKey0': 'Open',
  'StatusKey1': 'In Progress',
  'StatusKey2': 'Resolved',
  'StatusKey3': 'Closed'
} as const;
export type TicketStatusKey = keyof typeof TicketStatusKeyToLabel;

export interface Ticket {
  /**
   * @displayName Ticket
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Ticket Number
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  ticketNumber: string;
  /**
   * @displayName Assigned To
   */
  assignedTo?: string;
  /**
   * @displayName Attachment URL
   */
  attachmentURL?: string;
  /**
   * @displayName Company
   */
  company?: Pick<Company, 'id' | 'name1'>;
  /**
   * @displayName Created By
   * @validationRule Required for create/update operations
   */
  createdBy: string;
  /**
   * @displayName Created Date
   * @validationRule Required for create/update operations
   */
  createdDate: string;
  /**
   * @displayName Department
   */
  department?: Pick<Department, 'id' | 'name1'>;
  /**
   * @displayName Description
   * @validationRule Required for create/update operations
   */
  description: string;
  /**
   * @displayName Item Detail
   * @validationRule Required for create/update operations
   */
  itemDetail: string;
  /**
   * @displayName Main Category
   * @validationRule Required for create/update operations
   */
  mainCategory: string;
  /**
   * @displayName Mobile Number
   */
  mobileNumber?: string;
  /**
   * @displayName Priority
   * @validationRule Required for create/update operations
   */
  priorityKey: TicketPriorityKey;
  /**
   * @displayName Raised By Name
   */
  raisedByName?: string;
  /**
   * @displayName Requester Email
   */
  requesterEmail?: string;
  /**
   * @displayName SLA Due Date
   */
  sLADueDate?: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: TicketStatusKey;
  /**
   * @displayName Subcategory
   * @validationRule Required for create/update operations
   */
  subcategory: string;
  /**
   * @displayName Title
   * @validationRule Required for create/update operations
   */
  title: string;
}

export const _Ticket = 'Ticket' as const;